pyproject.toml
[tool.poetry]
name = "bangazon"
version = "0.1.0"
description = ""
authors = ["Björn Skarhaald <bjorn@valhalla.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
autopep8 = "^2.0.4"
astroid = "^3.1.0"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.14.0"
pylint = "^3.1.0"
Django = "^5.0.3"
pillow = "^10.2.0"
wheel = "^0.43.0"
pylint-django = "^2.5.5"
django-safedelete = "^1.3.3"
isort = "^5.13.2"
colorama = "^0.4.6"
lazy-object-proxy = "^1.10.0"
mccabe = "^0.7.0"
pycodestyle = "^2.11.1"
six = "^1.16.0"
setuptools = "^78.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
